extract an interface for commute time probability, and make it dependent on mode (although the default implementation is insensitive to this)
