+++ siloCore/src/main/java/de/tum/bgu/msm/models/relocation/moves/CarAndTransitHousingStrategyImpl.java
@@ -129,15 +129,15 @@ public class CarAndTransitHousingStrategyImpl implements HousingStrategy {
                 final Job job = jobDataManager.getJobFromId(pp.getJobId());
                 if(carToWorkersRatio == 0.) {
                     int ptTime = (int) travelTimes.getTravelTime(dwelling, job, properties.transportModel.peakHour_s, TransportMode.pt);
-                    factorForThisZone = commutingTimeProbability.getCommutingTimeProbability(Math.max(1, ptTime));
+                    factorForThisZone = commutingTimeProbability.getCommutingTimeProbability(Math.max(1, ptTime), TransportMode.pt);
                 } else if( carToWorkersRatio == 1.) {
                     int carTime = (int) travelTimes.getTravelTime(dwelling, job,  properties.transportModel.peakHour_s, TransportMode.car);
-                    factorForThisZone = commutingTimeProbability.getCommutingTimeProbability(Math.max(1, carTime));
+                    factorForThisZone = commutingTimeProbability.getCommutingTimeProbability(Math.max(1, carTime), TransportMode.car);
                 } else {
                     int carTime = (int) travelTimes.getTravelTime(dwelling, job,  properties.transportModel.peakHour_s, TransportMode.car);
                     int ptTime = (int) travelTimes.getTravelTime(dwelling, job,  properties.transportModel.peakHour_s, TransportMode.pt);
-                    double factorCar = commutingTimeProbability.getCommutingTimeProbability(Math.max(1, carTime));
-                    double factorPt = commutingTimeProbability.getCommutingTimeProbability(Math.max(1, ptTime));
+                    double factorCar = commutingTimeProbability.getCommutingTimeProbability(Math.max(1, carTime), TransportMode.car);
+                    double factorPt = commutingTimeProbability.getCommutingTimeProbability(Math.max(1, ptTime), TransportMode.pt);
                     factorForThisZone= factorCar * carToWorkersRatio + (1 - carToWorkersRatio) * factorPt;
                 }
                 workDistanceUtility *= factorForThisZone;
@@ -178,15 +178,15 @@ public class CarAndTransitHousingStrategyImpl implements HousingStrategy {
                     Zone workZone = geoData.getZones().get(job.getZoneId());
                     if(carToWorkersRatio <= 0.) {
                         int ptTime = (int) travelTimes.getTravelTimeFromRegion(region, workZone, properties.transportModel.peakHour_s, TransportMode.pt);
-                        thisRegionFactor = commutingTimeProbability.getCommutingTimeProbability(Math.max(1, ptTime));
+                        thisRegionFactor = commutingTimeProbability.getCommutingTimeProbability(Math.max(1, ptTime), TransportMode.pt);
                     } else if( carToWorkersRatio >= 1.) {
                         int carTime = (int) travelTimes.getTravelTimeFromRegion(region, workZone,properties.transportModel.peakHour_s, TransportMode.car);
-                        thisRegionFactor = commutingTimeProbability.getCommutingTimeProbability(Math.max(1, carTime));
+                        thisRegionFactor = commutingTimeProbability.getCommutingTimeProbability(Math.max(1, carTime), TransportMode.car);
                     } else {
                         int carTime = (int) travelTimes.getTravelTimeFromRegion(region, workZone, properties.transportModel.peakHour_s, TransportMode.car);
                         int ptTime = (int) travelTimes.getTravelTimeFromRegion(region, workZone, properties.transportModel.peakHour_s, TransportMode.pt);
-                        double factorCar = commutingTimeProbability.getCommutingTimeProbability(Math.max(1, carTime));
-                        double factorPt = commutingTimeProbability.getCommutingTimeProbability(Math.max(1, ptTime));
+                        double factorCar = commutingTimeProbability.getCommutingTimeProbability(Math.max(1, carTime), TransportMode.car);
+                        double factorPt = commutingTimeProbability.getCommutingTimeProbability(Math.max(1, ptTime),TransportMode.pt);
 
                         thisRegionFactor= factorCar * carToWorkersRatio + (1 - carToWorkersRatio) * factorPt;
                     }
